+++ okf/lib/okf/cli.rb
@@ -42,6 +42,7 @@ module OKF
       "concepts" => %w[id title type description tags generated_at generated_by generated trust status
                        stale_after sources backlog_ref dir top_dir links_out links_in],
       "files" => %w[path id dir type title description],
+      "references" => %w[path dir kind referenced_by],
       "directories" => %w[dir ancestor index_path present synthesized count types tags subdirs body listing],
       "dirs" => %w[dir ancestor count subtree subdirs],
       "bundles" => %w[slug title dir mount default missing]
@@ -506,6 +507,7 @@ require "okf/cli/stats"
 require "okf/cli/types"
 require "okf/cli/tags"
 require "okf/cli/files"
+require "okf/cli/references"
 require "okf/cli/catalog"
 require "okf/cli/graph"